Fix: Resolve critical reward calculation and transaction hash issues
CRITICAL FIXES:
- Fix duplicate reward accumulation across nodes (processed_emission_macroblocks tracking)
- Fix claim_rewards transaction hash validation error (calculate hash before submit)
- Fix halving and Phase 2 countdown display (use real Genesis timestamp)

REWARD SYSTEM (v2.90):
- Add processed_emission_macroblocks HashSet to prevent duplicate reward processing
- Implement RocksDB persistence for processed MacroBlocks
- Add restore_pending_reward() to recover rewards after node restart
- Fix inconsistent reward amounts when querying different nodes

TRANSACTION HASH ARCHITECTURE:
- Clarify BLAKE3 usage for transaction hashing (tx.calculate_hash)
- Fix handle_claim_rewards to set tx.hash before submission
- Update outdated comments about SHA3 vs BLAKE3
- Document P2P deduplication using SHA3(bincode) separately

ECONOMIC DISPLAY:
- Move economic logs to node.rs for proper architectural placement
- Add 4-hour periodic economic status display with reward claiming
- Show halving cycle, years until next halving, and Phase 2 countdown
- Remove duplicate halving logic from qnet-node.rs

EXPLORER SYNC:
- Reduce sync batch size from 500 to 100 blocks for stability
- Reduce parallel fetches from 500 to 100 for better resource management
- Change nonce column type from BIGINT to NUMERIC(20,0) for large values
- Add proper timeout handling for large blocks (60s for Genesis)

GITIGNORE:
- Add genesis_*_data/ directories to prevent committing blockchain data
- Add PostgreSQL dump patterns (*.dump, *.sql.gz, *_backup.sql)
- Add pgdata/ and postgres_data/ directories

STORAGE:
- Add save/load functions for processed_emission_macroblocks
- Integrate persistence in node startup and MacroBlock processing

This commit resolves the critical issues where:
1. Nodes showed different reward amounts due to duplicate processing
2. Claim rewards failed with 'Invalid transaction hash' error
3. Economic information was not displayed at startup
4. Blockchain data could accidentally be committed to git

+++ applications/qnet-explorer/frontend/migrations/001_init.sql
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS transactions (
     from_address TEXT NOT NULL,
     to_address TEXT,
     amount BIGINT NOT NULL,
-    nonce BIGINT NOT NULL,
+    nonce NUMERIC(20,0) NOT NULL,
     block BIGINT NOT NULL,
     timestamp BIGINT NOT NULL,
     gas_price BIGINT NOT NULL DEFAULT 0,
